Ready for release
There was still some bugs that make the application crashed when used
with the password protection.

I also added a banner at start with the version, as I tend to forget to
increment the version I simply used the commit hash.
Obviously the commit hash will always be one version behind but this
should do the trick for now.